chore: move getMaxZoomLevel function into /src/utils/bivariate
